requirements.txt
dash==2.8.0
dash_bootstrap_components==1.3.0
numpy==1.22.1
pandas==1.4.2
plotly==5.13.0
python_igraph==0.10.4
scikit_learn==1.2.2
